blog: testing Youtube component
RE: #25

If you use `.mdx` files instead of `.md` files you can leverage
react components directly in your markdown file.

Which _appears_ to be the easiest method of providing a method
of embedding a youtube video with responsive dimensions.

I've moved the 'floating-input-widgets' blog post to an `.mdx`
and included a test comonent directly in the file.  It appears to
work fine.

TODO: move the component to an actual component file and register
it in the global scope of the site to be available in every `/mdx`
file.